Add the functionality of the currency conversion

+++ Interfaces/ICurrencyService.cs
@@ -11,7 +11,7 @@
         /// <param name="fromCurrency">The currency to convert from.</param>
         /// <param name="toCurrency">The currency to convert to.</param>
         /// <returns>Returns the exchange rate between the two currencies.</returns>
-        Task<decimal> GetExchangeRateAsync(string fromCurrency, string toCurrency);
+        Task<double> GetExchangeRateAsync(string fromCurrency, string toCurrency);
         /// <summary>
         /// This method converts an amount from one currency to another.
         /// </summary>
@@ -19,6 +19,6 @@
         /// <param name="fromCurrency">the currency to convert from.</param>
         /// <param name="toCurrency">The currency to convert to.</param>
         /// <returns>Returns the converted amount.</returns>
-        Task<decimal> ConvertAmountAsync(decimal amount, string fromCurrency, string toCurrency);
+        Task<double> ConvertAmountAsync(double amount, string fromCurrency, string toCurrency);
     }
 }
